# Optimized ReID inference on GPU